What:将Carsousel组件形成独立文件
Why:

How:
