+++ frontend/src/Carousel/Carousel.css
@@ -1,0 +1,97 @@
+.carousel{
+    position: relative;
+    height: 100%;
+    width: 100%;
+    overflow: hidden;
+  }
+  .carousel__innerwrap{
+    width: 300%;
+    /* height: 100%; */
+    overflow-x: scroll;
+    display:flex; 
+    position:absolute;
+    top:0;
+    bottom: 0;
+    left:0;
+    overflow: hidden;
+    transition-property: left;
+    transition-duration: .5s;
+    transition-timing-function:ease-in;
+    transition-delay:.5s;
+  }
+  .carousel__box{
+    position: relative;
+    width: 50%;
+    height: 100%;
+    background-color: #FAFAFA;
+    overflow: hidden;
+    text-align: center;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  .carousel__img{
+    width: 1440px;
+    line-height: 0;
+    scale: 0.5;
+    transform: translateY(-50%);
+  }
+  .carousel__img--or2{
+    scale: 1;
+    transform: translateY(0);
+  }
+  .carousel__img--or3{
+    scale: 1.5;
+    transform: translateY(15%);
+  }
+  .carousel__text{
+    width: 100%;
+    background: transparent;
+    color: #fff;
+    position: relative;
+    z-index: 1;
+    padding: 100px 0 72px 0;
+    color: #111111;
+  }
+  .bg--black{
+    background-color: #111111;
+  }
+  .bg--gray{
+    background-color: #F1F1F3;
+  }
+  
+  .color--white{
+    color: #fff;
+  }
+  .carousel__slider{
+    position: absolute;
+    width: 100%;
+    text-align: center;
+    justify-content: center;
+    display: flex;
+    bottom: 50px;
+  }
+  .sliderbar{
+    width: 50px;
+    height: 2px;
+    border-radius: 2px;
+    border: 1px solid #979797;
+    position: relative;
+    margin-left: 5px;
+    overflow: hidden;
+    background-color: #A9A9A9;
+  }
+  
+  
+  .slider--active{
+    width: 0;
+    height: 100%;
+    top:10px;
+    left: -100%;
+    background-color: #fff;
+    z-index:1 ;
+    transition-property: width;
+    transition-duration: 3s;
+    transition-timing-function:ease-in;
+    transition-delay:.5s;
+  }
